Create quarkus specific single profile config resolver
